#2963 Rename enum to integer conversion related tests package.
