+++ processor/src/test/java/org/mapstruct/ap/test/conversion/_enum/EnumToIntegerEnum.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.conversion.enumtointeger;
+package org.mapstruct.ap.test.conversion._enum;
 
 public enum EnumToIntegerEnum {
     ARBITRARY_VALUE_ZERO,
