+++ processor/src/test/java/org/mapstruct/ap/test/conversion/_enum/EnumToIntegerTarget.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.conversion.enumtointeger;
+package org.mapstruct.ap.test.conversion._enum;
 
 public class EnumToIntegerTarget {
     private Integer enumValue;
